# external-site/internal/application/adverts/gen/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package gen

import (
	"fmt"
	"time"
)

func (s *ErrorStatusCode) Error() string {
	return fmt.Sprintf("code %d: %+v", s.StatusCode, s.Response)
}

// Ref: #/components/schemas/AdvertDetail
type AdvertDetail struct {
	// Advert title.
	Title string `json:"title"`
	// Detailed description of the advert.
	Description string `json:"description"`
	// Price in cents.
	Price int `json:"price"`
	// Publication date in RFC3339 format.
	PubDate time.Time `json:"pubDate"`
}

// GetTitle returns the value of Title.
func (s *AdvertDetail) GetTitle() string {
	return s.Title
}

// GetDescription returns the value of Description.
func (s *AdvertDetail) GetDescription() string {
	return s.Description
}

// GetPrice returns the value of Price.
func (s *AdvertDetail) GetPrice() int {
	return s.Price
}

// GetPubDate returns the value of PubDate.
func (s *AdvertDetail) GetPubDate() time.Time {
	return s.PubDate
}

// SetTitle sets the value of Title.
func (s *AdvertDetail) SetTitle(val string) {
	s.Title = val
}

// SetDescription sets the value of Description.
func (s *AdvertDetail) SetDescription(val string) {
	s.Description = val
}

// SetPrice sets the value of Price.
func (s *AdvertDetail) SetPrice(val int) {
	s.Price = val
}

// SetPubDate sets the value of PubDate.
func (s *AdvertDetail) SetPubDate(val time.Time) {
	s.PubDate = val
}

func (*AdvertDetail) getAdvertRes() {}

// Ref: #/components/schemas/AdvertListItem
type AdvertListItem struct {
	// Advert title.
	Title string `json:"title"`
	// Relative URL to the advert.
	URL string `json:"url"`
}

// GetTitle returns the value of Title.
func (s *AdvertListItem) GetTitle() string {
	return s.Title
}

// GetURL returns the value of URL.
func (s *AdvertListItem) GetURL() string {
	return s.URL
}

// SetTitle sets the value of Title.
func (s *AdvertListItem) SetTitle(val string) {
	s.Title = val
}

// SetURL sets the value of URL.
func (s *AdvertListItem) SetURL(val string) {
	s.URL = val
}

// DeleteAdvertNoContent is response for DeleteAdvert operation.
type DeleteAdvertNoContent struct{}

// Ref: #/components/schemas/Error
type Error struct {
	// Error message.
	Error string `json:"error"`
	// Error details.
	Detail OptErrorDetail `json:"detail"`
}

// GetError returns the value of Error.
func (s *Error) GetError() string {
	return s.Error
}

// GetDetail returns the value of Detail.
func (s *Error) GetDetail() OptErrorDetail {
	return s.Detail
}

// SetError sets the value of Error.
func (s *Error) SetError(val string) {
	s.Error = val
}

// SetDetail sets the value of Detail.
func (s *Error) SetDetail(val OptErrorDetail) {
	s.Detail = val
}

func (*Error) getAdvertRes() {}

// Error details.
type ErrorDetail map[string]string

func (s *ErrorDetail) init() ErrorDetail {
	m := *s
	if m == nil {
		m = map[string]string{}
		*s = m
	}
	return m
}

// ErrorStatusCode wraps Error with StatusCode.
type ErrorStatusCode struct {
	StatusCode int
	Response   Error
}

// GetStatusCode returns the value of StatusCode.
func (s *ErrorStatusCode) GetStatusCode() int {
	return s.StatusCode
}

// GetResponse returns the value of Response.
func (s *ErrorStatusCode) GetResponse() Error {
	return s.Response
}

// SetStatusCode sets the value of StatusCode.
func (s *ErrorStatusCode) SetStatusCode(val int) {
	s.StatusCode = val
}

// SetResponse sets the value of Response.
func (s *ErrorStatusCode) SetResponse(val Error) {
	s.Response = val
}

// NewOptErrorDetail returns new OptErrorDetail with value set to v.
func NewOptErrorDetail(v ErrorDetail) OptErrorDetail {
	return OptErrorDetail{
		Value: v,
		Set:   true,
	}
}

// OptErrorDetail is optional ErrorDetail.
type OptErrorDetail struct {
	Value ErrorDetail
	Set   bool
}

// IsSet returns true if OptErrorDetail was set.
func (o OptErrorDetail) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptErrorDetail) Reset() {
	var v ErrorDetail
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptErrorDetail) SetTo(v ErrorDetail) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptErrorDetail) Get() (v ErrorDetail, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptErrorDetail) Or(d ErrorDetail) ErrorDetail {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// NewOptInt returns new OptInt with value set to v.
func NewOptInt(v int) OptInt {
	return OptInt{
		Value: v,
		Set:   true,
	}
}

// OptInt is optional int.
type OptInt struct {
	Value int
	Set   bool
}

// IsSet returns true if OptInt was set.
func (o OptInt) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptInt) Reset() {
	var v int
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptInt) SetTo(v int) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptInt) Get() (v int, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptInt) Or(d int) int {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// NewOptNilString returns new OptNilString with value set to v.
func NewOptNilString(v string) OptNilString {
	return OptNilString{
		Value: v,
		Set:   true,
	}
}

// OptNilString is optional nullable string.
type OptNilString struct {
	Value string
	Set   bool
	Null  bool
}

// IsSet returns true if OptNilString was set.
func (o OptNilString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptNilString) Reset() {
	var v string
	o.Value = v
	o.Set = false
	o.Null = false
}

// SetTo sets value to v.
func (o *OptNilString) SetTo(v string) {
	o.Set = true
	o.Null = false
	o.Value = v
}

// IsNull returns true if value is Null.
func (o OptNilString) IsNull() bool { return o.Null }

// SetToNull sets value to null.
func (o *OptNilString) SetToNull() {
	o.Set = true
	o.Null = true
	var v string
	o.Value = v
}

// IsEmpty returns true if the field was omitted from the payload (not Set and not Null).
func (o OptNilString) IsEmpty() bool {
	return !o.Set && !o.Null
}

// Get returns value and boolean that denotes whether value was set.
func (o OptNilString) Get() (v string, ok bool) {
	if o.Null {
		return v, false
	}
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptNilString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/components/schemas/SearchResponse
type SearchResponse struct {
	// Current page number.
	Page int `json:"page"`
	// Total number of adverts.
	Total int `json:"total"`
	// URL to next page or null if last page.
	NextPage OptNilString `json:"nextPage"`
	// List of adverts on current page.
	Adverts []AdvertListItem `json:"adverts"`
}

// GetPage returns the value of Page.
func (s *SearchResponse) GetPage() int {
	return s.Page
}

// GetTotal returns the value of Total.
func (s *SearchResponse) GetTotal() int {
	return s.Total
}

// GetNextPage returns the value of NextPage.
func (s *SearchResponse) GetNextPage() OptNilString {
	return s.NextPage
}

// GetAdverts returns the value of Adverts.
func (s *SearchResponse) GetAdverts() []AdvertListItem {
	return s.Adverts
}

// SetPage sets the value of Page.
func (s *SearchResponse) SetPage(val int) {
	s.Page = val
}

// SetTotal sets the value of Total.
func (s *SearchResponse) SetTotal(val int) {
	s.Total = val
}

// SetNextPage sets the value of NextPage.
func (s *SearchResponse) SetNextPage(val OptNilString) {
	s.NextPage = val
}

// SetAdverts sets the value of Adverts.
func (s *SearchResponse) SetAdverts(val []AdvertListItem) {
	s.Adverts = val
}

type UpsertAdvertCreated AdvertDetail

func (*UpsertAdvertCreated) upsertAdvertRes() {}

type UpsertAdvertOK AdvertDetail

func (*UpsertAdvertOK) upsertAdvertRes() {}